Move all wpcom files to composite-checkout-wpcom
