+++ packages/composite-checkout-wpcom/src/components/coupon.js
