+++ packages/composite-checkout-wpcom/src/components/index.js
