+++ packages/composite-checkout-wpcom/src/components/join-classes.js
